+++ shaders/utils/getSkyColor.glsl
@@ -3,7 +3,7 @@
 
 
 
-vec3 getSkyColor(vec3 viewDir) {
+vec3 getSkyColor(vec3 viewDir, const bool includeLightning) {
 	
 	#ifdef OVERWORLD
 		
@@ -39,6 +39,12 @@ vec3 getSkyColor(vec3 viewDir) {
 		skyColor = 1.0 - (skyColor - 1.0) * (skyColor - 1.0);
 		skyColor *= 1.2;
 		skyColor *= 0.3 + 0.7 * dayPercent;
+		
+		if (includeLightning) {
+			uint time = uint(frameTimeCounter * 8.0);
+			float amount = float(sin(time * 4u) > 0.0);
+			skyColor += LIGHTNING_BRIGHTNESS * 0.25 * float(lightningBoltPosition.w > 0.001) * amount;
+		}
 		
 		#ifdef OVERWORLD
 			float eyeBrightness = eyeBrightnessSmooth.y / 240.0;